+++ resources/[dg]/dg-jobs/typescript/src/types/sanddigging.d.ts
@@ -1,6 +1,9 @@
 declare namespace Sanddigging {
   type Config = {
-    specialItemChance: number;
+    specialItemChance: {
+      min: number;
+      max: number;
+    };
     specialItems: string[];
     vehicle: Vec4;
     spots: Vec3[];